requirements.txt
<<<<<<< HEAD
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
greenlet==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pydantic==2.9.1
pydantic_core==2.23.3
typing_extensions==4.12.2
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.2
SQLAlchemy==2.0.20
Werkzeug==2.2.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
alembic==1.10.3
pytz==2023.3
=======
alembic==1.13.2
altgraph==0.17.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.2
filelock==3.15.4
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.0
fonttools==4.53.0
future==0.18.2
h11==0.14.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
macholib==1.15.2
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.1
passlib==1.7.4
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.7
pyasn1==0.6.0
pydantic==2.0.3
pydantic_core==2.3.0  # Aligned with pydantic 2.0.3
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
scipy==1.13.1
six==1.15.0
sniffio==1.3.1
SQLAlchemy==1.4.49
starlette==0.38.4
typing_extensions==4.12.2
uvicorn==0.30.6
virtualenv==20.26.3
Werkzeug==3.0.4
zipp==3.19.2
>>>>>>> 6873958458b7c396f826c02bb2bd911cc5fba600
